server/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
PyMySQL==1.1.0
bcrypt==4.1.2
python-dotenv==1.0.1
email-validator==2.1.1
PyJWT==2.1.0
